minor refactoring in platform specific javafx code

+++ src/main/java/de/helicopter_vs_aliens/platform_specific/javafx/GameApplication.java
@@ -7,7 +7,6 @@ import de.helicopter_vs_aliens.control.events.EventFactory;
 import de.helicopter_vs_aliens.graphics.Graphics2DAdapter;
 import de.helicopter_vs_aliens.graphics.GraphicsAdapter;
 import de.helicopter_vs_aliens.graphics.GraphicsManager;
-import de.helicopter_vs_aliens.graphics.JavaFxAdapter;
 import de.helicopter_vs_aliens.gui.window.Window;
 import de.helicopter_vs_aliens.util.Colorations;
 import de.helicopter_vs_aliens.util.geometry.Dimension;
@@ -112,7 +111,7 @@ public class GameApplication extends Application
 
         GraphicsContext graphicsContext2D = canvas.getGraphicsContext2D();
         graphicsFxAdapter = new JavaFxAdapter(graphicsContext2D);
-        offImage = new BufferedImage((int) VIRTUAL_DIMENSION.getWidth(), (int) VIRTUAL_DIMENSION.getHeight(), BufferedImage.TYPE_INT_RGB);
+        offImage = new BufferedImage(VIRTUAL_DIMENSION.getWidth(), VIRTUAL_DIMENSION.getHeight(), BufferedImage.TYPE_INT_RGB);
         graphicsAdapter = Graphics2DAdapter.of(offImage);
         graphicsAdapter.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
